refactor: content_type and associated
