+++ config/config.exs
@@ -93,7 +93,7 @@ config :ex_audit,
     WraftDoc.BlockTemplates.BlockTemplate,
     WraftDoc.Document.Asset,
     WraftDoc.Document.Block,
-    WraftDoc.Document.ContentType,
+    WraftDoc.ContentTypes.ContentType,
     WraftDoc.Document.DataTemplate,
     WraftDoc.Document.Field,
     WraftDoc.Document.Instance,
