+++ test/support/factory.ex
@@ -17,14 +17,14 @@ defmodule WraftDoc.Factory do
   alias WraftDoc.AuthTokens.AuthToken
   alias WraftDoc.BlockTemplates.BlockTemplate
   alias WraftDoc.Comments.Comment
+  alias WraftDoc.ContentTypes.ContentType
+  alias WraftDoc.ContentTypes.ContentTypeField
+  alias WraftDoc.ContentTypes.ContentTypeRole
   alias WraftDoc.Document.Asset
   alias WraftDoc.Document.Block
   alias WraftDoc.Document.CollectionForm
   alias WraftDoc.Document.CollectionFormField
   alias WraftDoc.Document.ContentCollaboration
-  alias WraftDoc.Document.ContentType
-  alias WraftDoc.Document.ContentTypeField
-  alias WraftDoc.Document.ContentTypeRole
   alias WraftDoc.Document.Counter
   alias WraftDoc.Document.DataTemplate
   alias WraftDoc.Document.Engine
