+++ test/wraft_doc/document/content_type_field_test.exs
@@ -2,7 +2,7 @@ defmodule WraftDocTest.Document.ContentTypeFieldTest do
   @moduledoc false
   use WraftDoc.ModelCase
   @moduletag :document
-  alias WraftDoc.Document.ContentTypeField
+  alias WraftDoc.ContentTypes.ContentTypeField
   import WraftDoc.Factory
 
   @invalid_attrs %{content_type_id: nil, field_id: nil}
